+++ src/physics/large_scale_condensation.jl
@@ -3,24 +3,24 @@ abstract type AbstractCondensation <: AbstractParameterization end
 export ImplicitCondensation
 """Large-scale condensation with implicit latent heat release.
 $(TYPEDFIELDS)"""
-@kwdef struct ImplicitCondensation{NF} <: AbstractCondensation
+@parameterized @kwdef struct ImplicitCondensation{NF} <: AbstractCondensation
     "[OPTION] Relative humidity threshold [1 = 100%] to trigger condensation"
-    relative_humidity_threshold::NF = 0.95
+    @param relative_humidity_threshold::NF = 0.95 (bounds=0..1,)
 
     "[OPTION] Reevaporation efficiency [1/(kg/kg)], 0 for no reevaporation"
-    reevaporation::NF = 30
+    @param reevaporation::NF = 30 (bounds=Nonnegative,)
 
     "[OPTION] Convert rain below freezing to snow?"
     snow::Bool = true
 
     "[OPTION] Freezing temperature for snow fall [K]"
-    freezing_threshold::NF = 263
+    @param freezing_threshold::NF = 263 (bounds=Positive,)
 
     "[OPTION] Melting temperature for snow fall [K]"
-    melting_threshold::NF = 278
+    @param melting_threshold::NF = 278 (bounds=Positive,)
 
     "[OPTION] Time scale in multiples of time step Δt, the larger the less immediate"
-    time_scale::NF = 3
+    @param time_scale::NF = 3 (bounds=Positive,)
 end
 
 Adapt.@adapt_structure ImplicitCondensation