+++ src/physics/radiation/shortwave_radiation.jl
@@ -71,11 +71,12 @@ based on relative humidity and precipitation, with radiative transfer through cl
 Cloud cover is calculated as a combination of relative humidity and precipitation contributions,
 and a cloud albedo is applied to the downward beam. Fields and options are
 
-$(TYPEDFIELDS)"""
-struct OneBandShortwave{C, T, R} <: AbstractShortwave
-    clouds::C
-    transmissivity::T
-    radiative_transfer::R
+$(TYPEDFIELDS)
+"""
+@parameterized @kwdef struct OneBandShortwave{C, T, R} <: AbstractShortwave
+    @component clouds::C
+    @component transmittance::T
+    @component radiative_transfer::R
 end
 
 # primitive wet model version
@@ -134,12 +135,12 @@ function shortwave_radiation!(
 end
 
 export OneBandShortwaveRadiativeTransfer
-@kwdef struct OneBandShortwaveRadiativeTransfer{NF} <: AbstractShortwaveRadiativeTransfer
+@parameterized @kwdef struct OneBandShortwaveRadiativeTransfer{NF} <: AbstractShortwaveRadiativeTransfer
     "[OPTION] Ozone absorption in upper stratosphere (W/m^2)"
-    ozone_absorp_upper::NF = 0
+    @param ozone_absorp_upper::NF = 0 (bounds=Nonnegative,)
 
     "[OPTION] Ozone absorption in lower stratosphere (W/m^2)"
-    ozone_absorp_lower::NF = 0
+    @param ozone_absorp_lower::NF = 0 (bounds=Nonnegative,)
 end
 
 # generator function